Add files via upload
This is a mini-project using Node JS and MYSQL. 

I used Node JS to connect to my local Mysql server and made a website. 
I used ejs file to design the structure of the website

What this website does is that stores the email of users and renews automatically when someone signs up.